move stepcard into own component, add zeroswap assets
