+++ components/Modals/BuyWithCoinbaseModal.tsx
@@ -1,6 +1,5 @@
 import { sequence } from "0xsequence";
 import { polygon } from "@wagmi/chains";
-import clsx from "clsx";
 import { BigNumber, utils } from "ethers";
 import Image from "next/image";
 import { useContext, useState, useEffect } from "react";
@@ -8,6 +7,7 @@ import { Tooltip } from "react-tooltip";
 import { useAccount, useBalance, useNetwork, useSwitchNetwork } from "wagmi";
 
 import PaymentOptionModal from "@/components/Modals/PaymentOptionModal";
+import StepCard from "@/components/Modals/StepCard";
 import { ModalContext } from "@/lib/context";
 import { sliceAddress } from "@/lib/utils";
 import { buyWithSwap } from "@/payments";
@@ -45,7 +45,6 @@ export default function BuyWithCoinbaseModal({
 
   const userIsOnPolygon = chain?.id === polygon.id;
   const isSequenceWallet = connector?.id === "sequence";
-  const isMetamaskWallet = connector?.id === "metaMask";
 
   useEffect(() => {
     if (isCopiedTooltipOpen) {
@@ -71,93 +70,6 @@ export default function BuyWithCoinbaseModal({
       if (val.gte(currBuyAmt)) setIsProviderStepDone(true);
     }
   }, [balance]);
-
-  const StepCard = ({
-    index,
-    iconPath,
-    title,
-    content,
-    action,
-    done = false,
-  }: {
-    index: number;
-    iconPath: string;
-    title: string;
-    content: string;
-    action: any;
-    done?: boolean;
-  }) => (
-    <div
-      className={clsx(
-        "cursor-pointer flex flex-col justify-center border-2 rounded-xl border-pine-100 hover:border-pine-400 mb-2",
-        done && "bg-cyan-600/20"
-      )}
-      onClick={action}
-    >
-      <div className="flex flex-col justify-center">
-        <div className="flex items-center p-3">
-          <div
-            className={clsx(
-              "relative circle border-2 w-[32px] h-[32px]",
-              done && "border-none bg-cyan-600 w-[32px] h-[32px]"
-            )}
-          >
-            {!done ? (
-              index
-            ) : (
-              <Image
-                alt="checkmark"
-                src="check-alpha.svg"
-                height={12}
-                width={12}
-              />
-            )}
-            <div
-              className={clsx(
-                "circle w-[20px] h-[20px] absolute top-[-7px] right-[-10px]",
-                done && "top-[-5px] right-[-8px]"
-              )}
-            >
-              <Image alt={iconPath} src={iconPath} height={20} width={20} />
-            </div>
-          </div>
-          <div className="pl-4">
-            <h5 className="text-sm mb-2">{title}</h5>
-            <p className="copy text-xs">
-              {content}{" "}
-              {index === 3 && isSequenceWallet && (
-                <>
-                  <Image
-                    alt="qrcode"
-                    style={{ display: "inline" }}
-                    src="/miniqr.svg"
-                    height={16}
-                    width={16}
-                  />{" "}
-                  +&nbsp;
-                  <Image
-                    alt="copypaste"
-                    style={{ display: "inline" }}
-                    src="/copy.svg"
-                    height={16}
-                    width={16}
-                  />
-                </>
-              )}
-            </p>
-          </div>
-        </div>
-        {index === 2 && (
-          <div className="p-3 border-t-2 flex justify-center w-full">
-            <Image alt="usdc" src="usdc.svg" height={20} width={20} />
-            <span className="ml-2 copy text-pine-900 font-bold">
-              Current USDC balance: {USDC}
-            </span>
-          </div>
-        )}
-      </div>
-    </div>
-  );
 
   return (
     <div className="flex flex-col max-w-[343px] text-pine-900 p-2">
@@ -189,7 +101,9 @@ export default function BuyWithCoinbaseModal({
         </button>
       </div>
       <section className="text-center">
-        <h3 className="pt-0">Buying Glo Dollars through {provider} and DEX</h3>
+        <h3 className="pt-0">
+          Buying Glo Dollars through {provider} and Uniswap
+        </h3>
         <p className="text-sm py-6">
           You can get Glo Dollars by exchanging another stablecoin called{" "}
           <b>USDC</b> for Glo Dollar using the <b>Uniswap</b> app.
@@ -215,6 +129,7 @@ export default function BuyWithCoinbaseModal({
             buyWithProvider();
           }}
           done={isProviderStepDone}
+          USDC={USDC}
         />
         <StepCard
           index={3}
@@ -230,6 +145,7 @@ export default function BuyWithCoinbaseModal({
             setIsSwapStepDone(true);
           }}
           done={isSwapStepDone}
+          isSequenceWallet={isSequenceWallet}
         />
         {isSequenceWallet && (
           <StepCard
